+++ e3sm_diags/derivations/derivations.py
@@ -1,656 +1,64 @@
 """
 This module defines functions for deriving variables using other variables.
+
+`derived_variables` is a dictionary to accomodate user specified derived
+variables. The driver uses the Dataset class to search for available variable
+keys and attempts to map them to a formula function to calculate a derived
+variable.
+
+For example to derived 'PRECT':
+  1. In `derived_variable` there is an entry for 'PRECT'.
+  2. The netCDF file does not have a 'PRECT' variable, but has 'PRECC' and
+     'PRECT'
+  3. To derived 'PRECT', 'PRECC' and 'PRECL' are passed as arguments to the
+     formula function 'prect()'.
 """
 from collections import OrderedDict
-from typing import TYPE_CHECKING, Callable, Dict, Optional, Tuple
-
-import MV2
-import numpy as np
-import xarray as xr
-from genutil import udunits
-
-if TYPE_CHECKING:
-    from cdms2.axis import FileAxis
-    from cdms2.fvariable import FileVariable
-
-AVOGADOR_CONS = 6.022e23
-
-
-def rename(new_name):
-    """Given the new name, just return it."""
-    return new_name
-
-
-def aplusb(var1, var2, target_units=None):
-    """Returns var1 + var2. If both of their units are not the same,
-    it tries to convert both of their units to target_units"""
-
-    if target_units is not None:
-        var1 = convert_units(var1, target_units)
-        var2 = convert_units(var2, target_units)
-
-    return var1 + var2
-
-
-def convert_units(var: xr.DataArray, target_units: str):  # noqa: C901
-    if var.attrs.get("units") is None:
-        if var.name == "SST":
-            var.attrs["units"] = target_units
-        elif var.name == "ICEFRAC":
-            var.attrs["units"] = target_units
-            var = 100.0 * var
-        elif var.name == "AODVIS":
-            var.attrs["units"] = target_units
-        elif var.name == "AODDUST":
-            var.attrs["units"] = target_units
-    elif var.name == "FAREA_BURNED":
-        var = var * 1e9
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "gC/m^2":
-        var = var / 1000.0
-        var.attrs["units"] = target_units
-    elif var.name == "FLOODPLAIN_VOLUME" and target_units == "km3":
-        var = var / 1.0e9
-        var.attrs["units"] = target_units
-    elif var.name == "AOD_550_ann":
-        var.attrs["units"] = target_units
-    elif var.name == "AOD_550":
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "C" and target_units == "DegC":
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "N/m2" and target_units == "N/m^2":
-        var.attrs["units"] = target_units
-    elif var.name == "AODVIS" or var.name == "AOD_550_ann" or var.name == "TOTEXTTAU":
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "fraction":
-        var = 100.0 * var
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "mb":
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "gpm":  # geopotential meter
-        var = var / 9.8 / 100  # convert to hecto meter
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "Pa/s":
-        var = var / 100.0 * 24 * 3600
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] == "mb/day":
-        var = var
-        var.attrs["units"] = target_units
-    elif var.name == "prw" and var.attrs["units"] == "cm":
-        var = var * 10.0  # convert from 'cm' to 'kg/m2' or 'mm'
-        var.attrs["units"] = target_units
-    elif var.attrs["units"] in ["gC/m^2/s"] and target_units == "g*/m^2/day":
-        var = var * 24 * 3600
-        var.attrs["units"] = var.attrs["units"][0:7] + "day"
-    elif (
-        var.attrs["units"] in ["gN/m^2/s", "gP/m^2/s"] and target_units == "mg*/m^2/day"
-    ):
-        var = var * 24 * 3600 * 1000.0
-        var.attrs["units"] = "m" + var.attrs["units"][0:7] + "day"
-    elif var.attrs["units"] in ["gN/m^2/day", "gP/m^2/day", "gC/m^2/day"]:
-        pass
-    else:
-        temp = udunits(1.0, var.attrs["units"])
-        coeff, offset = temp.how(target_units)
-        var = coeff * var + offset
-        var.attrs["units"] = target_units
-
-    return var
-
-
-def _apply_land_sea_mask(
-    var: xr.DataArray, var_mask: xr.DataArray, lower_limit: float
-) -> xr.DataArray:
-    """Apply a land or sea mask on the variable.
-
-    Parameters
-    ----------
-    var : xr.DataArray
-        The variable.
-    var_mask : xr.DataArray
-        The variable mask ("LANDFRAC" or "OCNFRAC").
-    lower_limit : float
-        Update the mask variable with a lower limit. All values below the
-        lower limit will be masked.
-
-    Returns
-    -------
-    xr.DataArray
-        The masked variable.
-    """
-    cond = var_mask > lower_limit
-    masked_var = var.where(cond=cond, drop=False)
-
-    return masked_var
-
-
-def qflxconvert_units(var):
-    if (
-        var.attrs["units"] == "kg/m2/s"
-        or var.attrs["units"] == "kg m-2 s-1"
-        or var.attrs["units"] == "mm/s"
-    ):
-        # need to find a solution for units not included in udunits
-        # var = convert_units( var, 'kg/m2/s' )
-        var = var * 3600.0 * 24  # convert to mm/day
-        var.attrs["units"] = "mm/day"
-    elif var.attrs["units"] == "mm/hr":
-        var = var * 24.0
-        var.attrs["units"] = "mm/day"
-    return var
-
-
-def w_convert_q(var):
-    if var.attrs["units"] == "mol/mol":
-        var = (
-            var * 18.0 / 28.97 * 1000.0
-        )  # convert from volume mixing ratio to mass mixing ratio in units g/kg
-        var.attrs["units"] = "g/kg"
-        var.attrs["long_name"] = "H2OLNZ (radiation)"
-    return var
-
-
-def molec_convert_units(var, molar_weight):
-    # Convert molec/cm2/s to kg/m2/s
-    if var.attrs["units"] == "molec/cm2/s":
-        var = var / AVOGADOR_CONS * molar_weight * 10.0
-        var.attrs["units"] == "kg/m2/s"
-    return var
-
-
-def qflx_convert_to_lhflx(qflx, precc, precl, precsc, precsl):
-    # A more precise formula to close atmospheric energy budget:
-    # LHFLX is modified to account for the latent energy of frozen precipitation.
-    # LHFLX = (Lv+Lf)*QFLX - Lf*1.e3*(PRECC+PRECL-PRECSC-PRECSL)
-    # Constants, from AMWG diagnostics
-    Lv = 2.501e6
-    Lf = 3.337e5
-    var = (Lv + Lf) * qflx - Lf * 1.0e3 * (precc + precl - precsc - precsl)
-    var.attrs["units"] = "W/m2"
-    var.attrs["long_name"] = "Surface latent heat flux"
-    return var
-
-
-def qflx_convert_to_lhflx_approxi(var):
-    # QFLX units: kg/((m^2)*s)
-    # Multiply by the latent heat of condensation/vaporization (in J/kg)
-    # kg/((m^2)*s) * J/kg = J/((m^2)*s) = (W*s)/((m^2)*s) = W/(m^2)
-    var.attrs["units"] = "W/m2"
-    return var * 2.5e6
-
-
-def pminuse_convert_units(var):
-    if (
-        var.attrs["units"] == "kg/m2/s"
-        or var.attrs["units"] == "kg m-2 s-1"
-        or var.attrs["units"] == "kg/s/m^2"
-    ):
-        # need to find a solution for units not included in udunits
-        # var = convert_units( var, 'kg/m2/s' )
-        var = var * 3600.0 * 24  # convert to mm/day
-        var.attrs["units"] = "mm/day"
-    var.attrs["long_name"] = "precip. flux - evap. flux"
-    return var
-
-
-def prect(precc, precl):
-    """Total precipitation flux = convective + large-scale"""
-    with xr.set_options(keep_attrs=True):
-        var = precc + precl
-
-    var = convert_units(var, "mm/day")
-    var.name = "PRECT"
-    var.attrs["long_name"] = "Total precipitation rate (convective + large-scale)"
-    return var
-
-
-def precst(precc, precl):
-    """Total precipitation flux = convective + large-scale"""
-    with xr.set_options(keep_attrs=True):
-        var = precc + precl
-
-    var = convert_units(var, "mm/day")
-    var.name = "PRECST"
-    var.attrs["long_name"] = "Total snowfall flux (convective + large-scale)"
-    return var
-
-
-def tref_range(tmax, tmin):
-    """TREF daily range = TREFMXAV - TREFMNAV"""
-    var = tmax - tmin
-    var.name = "TREF_range"
-    var.attrs["units"] = "K"
-    var.attrs["long_name"] = "Surface Temperature Daily Range"
-    return var
-
-
-def tauxy(taux, tauy):
-    """tauxy = (taux^2 + tauy^2)sqrt"""
-    with xr.set_options(keep_attrs=True):
-        var = (taux**2 + tauy**2) ** 0.5
-
-    var = convert_units(var, "N/m^2")
-    var.name = "TAUXY"
-    var.attrs["long_name"] = "Total surface wind stress"
-    return var
-
-
-def fp_uptake(a, b):
-    """plant uptake of soil mineral N"""
-    var = a / b
-    var.name = "FP_UPTAKE"
-    var.attrs["units"] = "dimensionless"
-    var.attrs["long_name"] = "Plant uptake of soil mineral N"
-    return var
-
-
-def albedo(rsdt, rsut):
-    """TOA (top-of-atmosphere) albedo, rsut / rsdt, unit is nondimension"""
-    var = rsut / rsdt
-    var.name = "ALBEDO"
-    var.attrs["units"] = "dimensionless"
-    var.attrs["long_name"] = "TOA albedo"
-    return var
-
-
-def albedoc(rsdt, rsutcs):
-    """TOA (top-of-atmosphere) albedo clear-sky, rsutcs / rsdt, unit is nondimension"""
-    var = rsutcs / rsdt
-    var.name = "ALBEDOC"
-    var.attrs["units"] = "dimensionless"
-    var.attrs["long_name"] = "TOA albedo clear-sky"
-    return var
-
-
-def albedo_srf(rsds, rsus):
-    """Surface albedo, rsus / rsds, unit is nondimension"""
-    var = rsus / rsds
-    var.name = "ALBEDOC_SRF"
-    var.attrs["units"] = "dimensionless"
-    var.attrs["long_name"] = "Surface albedo"
-    return var
-
-
-def rst(rsdt, rsut):
-    """TOA (top-of-atmosphere) net shortwave flux"""
-    with xr.set_options(keep_attrs=True):
-        var = rsdt - rsut
-
-    var.name = "FSNTOA"
-    var.attrs["long_name"] = "TOA net shortwave flux"
-    return var
-
-
-def rstcs(rsdt, rsutcs):
-    """TOA (top-of-atmosphere) net shortwave flux clear-sky"""
-    with xr.set_options(keep_attrs=True):
-        var = rsdt - rsutcs
-
-    var.name = "FSNTOAC"
-    var.attrs["long_name"] = "TOA net shortwave flux clear-sky"
-    return var
-
-
-def swcfsrf(fsns, fsnsc):
-    """Surface shortwave cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = fsns - fsnsc
-
-    var.name = "SCWFSRF"
-    var.attrs["long_name"] = "Surface shortwave cloud forcing"
-    return var
-
-
-def lwcfsrf(flns, flnsc):
-    """Surface longwave cloud forcing, for ACME model, upward is postitive for LW , for ceres, downward is postive for both LW and SW"""
-    with xr.set_options(keep_attrs=True):
-        var = -(flns - flnsc)
-
-    var.name = "LCWFSRF"
-    var.attrs["long_name"] = "Surface longwave cloud forcing"
-    return var
-
-
-def swcf(fsntoa, fsntoac):
-    """TOA shortwave cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = fsntoa - fsntoac
-
-    var.name = "SWCF"
-    var.attrs["long_name"] = "TOA shortwave cloud forcing"
-    return var
-
-
-def lwcf(flntoa, flntoac):
-    """TOA longwave cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = flntoa - flntoac
-
-    var.name = "LWCF"
-    var.attrs["long_name"] = "TOA longwave cloud forcing"
-    return var
-
-
-def netcf2(swcf, lwcf):
-    """TOA net cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = swcf + lwcf
-
-    var.name = "NETCF"
-    var.attrs["long_name"] = "TOA net cloud forcing"
-    return var
-
-
-def netcf4(fsntoa, fsntoac, flntoa, flntoac):
-    """TOA net cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = fsntoa - fsntoac + flntoa - flntoac
-
-    var.name = "NETCF"
-    var.attrs["long_name"] = "TOA net cloud forcing"
-    return var
-
-
-def netcf2srf(swcf, lwcf):
-    """Surface net cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = swcf + lwcf
-
-    var.name = "NETCF_SRF"
-    var.attrs["long_name"] = "Surface net cloud forcing"
-    return var
-
-
-def netcf4srf(fsntoa, fsntoac, flntoa, flntoac):
-    """Surface net cloud forcing"""
-    with xr.set_options(keep_attrs=True):
-        var = fsntoa - fsntoac + flntoa - flntoac
-
-    var.name = "NETCF4SRF"
-    var.attrs["long_name"] = "Surface net cloud forcing"
-    return var
-
-
-def fldsc(ts, flnsc):
-    """Clearsky Surf LW downwelling flux"""
-    with xr.set_options(keep_attrs=True):
-        var = 5.67e-8 * ts**4 - flnsc
-
-    var.name = "FLDSC"
-    var.attrs["units"] = "W/m2"
-    var.attrs["long_name"] = "Clearsky Surf LW downwelling flux"
-    return var
-
-
-def restom(fsnt, flnt):
-    """TOM(top of model) Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = fsnt - flnt
-
-    var.name = "RESTOM"
-    var.attrs["long_name"] = "TOM(top of model) Radiative flux"
-    return var
-
-
-def restoa(fsnt, flnt):
-    """TOA(top of atmosphere) Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = fsnt - flnt
-
-    var.name = "RESTOA"
-    var.attrs["long_name"] = "TOA(top of atmosphere) Radiative flux"
-    return var
-
-
-def flus(flds, flns):
-    """Surface Upwelling LW Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = flns + flds
-
-    var.name = "FLUS"
-    var.attrs["long_name"] = "Upwelling longwave flux at surface"
-    return var
-
-
-def fsus(fsds, fsns):
-    """Surface Up-welling SW Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = fsds - fsns
-
-    var.name = "FSUS"
-    var.attrs["long_name"] = "Upwelling shortwave flux at surface"
-    return var
-
-
-def netsw(rsds, rsus):
-    """Surface SW Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = rsds - rsus
-
-    var.name = "FSNS"
-    var.attrs["long_name"] = "Surface SW Radiative flux"
-    return var
-
-
-def netlw(rlds, rlus):
-    """Surface LW Radiative flux"""
-    with xr.set_options(keep_attrs=True):
-        var = -(rlds - rlus)
-
-    var.name = "NET_FLUX_SRF"
-    var.attrs["long_name"] = "Surface LW Radiative flux"
-    return var
-
-
-def netflux4(fsns, flns, lhflx, shflx):
-    """Surface Net flux"""
-    with xr.set_options(keep_attrs=True):
-        var = fsns - flns - lhflx - shflx
-
-    var.name = "NET_FLUX_SRF"
-    var.attrs["long_name"] = "Surface Net flux"
-    return var
-
-
-def netflux6(rsds, rsus, rlds, rlus, hfls, hfss):
-    """Surface Net flux"""
-    with xr.set_options(keep_attrs=True):
-        var = rsds - rsus + (rlds - rlus) - hfls - hfss
-
-    var.name = "NET_FLUX_SRF"
-    var.attrs["long_name"] = "Surface Net flux"
-    return var
-
-
-def adjust_prs_val_units(
-    prs: "FileAxis", prs_val: float, prs_val0: Optional[float]
-) -> float:
-    # FIXME: Refactor this function to operate on xr.Dataset/xr.DataArray.
-    """Adjust the prs_val units based on the prs.id"""
-    # COSP v2 cosp_pr in units Pa instead of hPa as in v1
-    # COSP v2 cosp_htmisr in units m instead of km as in v1
-    adjust_ids = {"cosp_prs": 100, "cosp_htmisr": 1000}
-
-    if prs_val0:
-        prs_val = prs_val0
-        if prs.id in adjust_ids.keys() and max(prs.getData()) > 1000:
-            prs_val = prs_val * adjust_ids[prs.id]
-
-    return prs_val
-
-
-def determine_cloud_level(
-    prs_low: float,
-    prs_high: float,
-    low_bnds: Tuple[int, int],
-    high_bnds: Tuple[int, int],
-) -> str:
-    """Determines the cloud type based on prs values and the specified boundaries"""
-    # Threshold for cloud top height: high cloud (<440hPa or > 7km), midlevel cloud (440-680hPa, 3-7 km) and low clouds (>680hPa, < 3km)
-    if prs_low in low_bnds and prs_high in high_bnds:
-        return "middle cloud fraction"
-    elif prs_low in low_bnds:
-        return "high cloud fraction"
-    elif prs_high in high_bnds:
-        return "low cloud fraction"
-    else:
-        return "total cloud fraction"
-
-
-def cosp_bin_sum(
-    cld: "FileVariable",
-    prs_low0: Optional[float],
-    prs_high0: Optional[float],
-    tau_low0: Optional[float],
-    tau_high0: Optional[float],
-):
-    # FIXME: Refactor this function to operate on xr.Dataset/xr.DataArray.
-    """sum of cosp bins to calculate cloud fraction in specified cloud top pressure / height and
-    cloud thickness bins, input variable has dimension (cosp_prs,cosp_tau,lat,lon)/(cosp_ht,cosp_tau,lat,lon)"""
-    prs: FileAxis = cld.getAxis(0)
-    tau: FileAxis = cld.getAxis(1)
-
-    prs_low: float = adjust_prs_val_units(prs, prs[0], prs_low0)
-    prs_high: float = adjust_prs_val_units(prs, prs[-1], prs_high0)
-
-    if prs_low0 is None and prs_high0 is None:
-        prs_lim = "total cloud fraction"
-
-    tau_high, tau_low, tau_lim = determine_tau(tau, tau_low0, tau_high0)
-
-    if cld.id == "FISCCP1_COSP":  # ISCCP model
-        cld_bin = cld(cosp_prs=(prs_low, prs_high), cosp_tau=(tau_low, tau_high))
-        simulator = "ISCCP"
-    if cld.id == "CLISCCP":  # ISCCP obs
-        cld_bin = cld(isccp_prs=(prs_low, prs_high), isccp_tau=(tau_low, tau_high))
-
-    if cld.id == "CLMODIS":  # MODIS
-        prs_lim = determine_cloud_level(prs_low, prs_high, (440, 44000), (680, 68000))
-        simulator = "MODIS"
-
-        if prs.id == "cosp_prs":  # Model
-            cld_bin = cld(
-                cosp_prs=(prs_low, prs_high), cosp_tau_modis=(tau_low, tau_high)
-            )
-        elif prs.id == "modis_prs":  # Obs
-            cld_bin = cld(modis_prs=(prs_low, prs_high), modis_tau=(tau_low, tau_high))
-
-    if cld.id == "CLD_MISR":  # MISR model
-        if max(prs) > 1000:  # COSP v2 cosp_htmisr in units m instead of km as in v1
-            cld = cld[
-                1:, :, :, :
-            ]  # COSP v2 cosp_htmisr[0] equals to 0 instead of -99 as in v1, therefore cld needs to be masked manually
-        cld_bin = cld(cosp_htmisr=(prs_low, prs_high), cosp_tau=(tau_low, tau_high))
-        prs_lim = determine_cloud_level(prs_low, prs_high, (7, 7000), (3, 3000))
-        simulator = "MISR"
-    if cld.id == "CLMISR":  # MISR obs
-        cld_bin = cld(misr_cth=(prs_low, prs_high), misr_tau=(tau_low, tau_high))
-
-    cld_bin_sum = MV2.sum(MV2.sum(cld_bin, axis=1), axis=0)
-
-    try:
-        cld_bin_sum.long_name = simulator + ": " + prs_lim + " with " + tau_lim
-        # cld_bin_sum.long_name = "{}: {} with {}".format(simulator, prs_lim, tau_lim)
-    except BaseException:
-        pass
-    return cld_bin_sum
-
-
-def determine_tau(
-    tau: "FileAxis", tau_low0: Optional[float], tau_high0: Optional[float]
-):
-    # FIXME: Refactor this function to operate on xr.Dataset/xr.DataArray.
-    tau_low = tau[0]
-    tau_high = tau[-1]
-
-    if tau_low0 is None and tau_high0:
-        tau_high = tau_high0
-        tau_lim = "tau <" + str(tau_high0)
-    elif tau_high0 is None and tau_low0:
-        tau_low = tau_low0
-        tau_lim = "tau >" + str(tau_low0)
-    elif tau_low0 is None and tau_high0 is None:
-        tau_lim = str(tau_low) + "< tau < " + str(tau_high)
-    else:
-        tau_low = tau_low0
-        tau_high = tau_high0
-        tau_lim = str(tau_low) + "< tau < " + str(tau_high)
-
-    return tau_high, tau_low, tau_lim
-
-
-def cosp_histogram_standardize(cld: "FileVariable"):
-    # FIXME: Refactor this function to operate on xr.Dataset/xr.DataArray.
-    """standarize cloud top pressure and cloud thickness bins to dimensions that
-    suitable for plotting, input variable has dimention (cosp_prs,cosp_tau)"""
-    prs = cld.getAxis(0)
-    tau = cld.getAxis(1)
-
-    prs[0]
-    prs_high = prs[-1]
-    tau[0]
-    tau_high = tau[-1]
-
-    prs_bounds = getattr(prs, "bounds")
-    if prs_bounds is None:
-        cloud_prs_bounds = np.array(
-            [
-                [1000.0, 800.0],
-                [800.0, 680.0],
-                [680.0, 560.0],
-                [560.0, 440.0],
-                [440.0, 310.0],
-                [310.0, 180.0],
-                [180.0, 0.0],
-            ]
-        )  # length 7
-        prs.setBounds(np.array(cloud_prs_bounds, dtype=np.float32))
-
-    tau_bounds = getattr(tau, "bounds")
-    if tau_bounds is None:
-        cloud_tau_bounds = np.array(
-            [
-                [0.3, 1.3],
-                [1.3, 3.6],
-                [3.6, 9.4],
-                [9.4, 23],
-                [23, 60],
-                [60, 379],
-            ]
-        )  # length 6
-        tau.setBounds(np.array(cloud_tau_bounds, dtype=np.float32))
-
-    if cld.id == "FISCCP1_COSP":  # ISCCP model
-        cld_hist = cld(cosp_tau=(0.3, tau_high))
-    if cld.id == "CLISCCP":  # ISCCP obs
-        cld_hist = cld(isccp_tau=(0.3, tau_high))
-
-    if cld.id == "CLMODIS":  # MODIS
-        try:
-            cld_hist = cld(cosp_tau_modis=(0.3, tau_high))  # MODIS model
-        except BaseException:
-            cld_hist = cld(modis_tau=(0.3, tau_high))  # MODIS obs
-
-    if cld.id == "CLD_MISR":  # MISR model
-        if max(prs) > 1000:  # COSP v2 cosp_htmisr in units m instead of km as in v1
-            cld = cld[
-                1:, :, :, :
-            ]  # COSP v2 cosp_htmisr[0] equals to 0 instead of -99 as in v1, therefore cld needs to be masked manually
-            prs_high = 1000.0 * prs_high
-        cld_hist = cld(cosp_tau=(0.3, tau_high), cosp_htmisr=(0, prs_high))
-    if cld.id == "CLMISR":  # MISR obs
-        cld_hist = cld(misr_tau=(0.3, tau_high), misr_cth=(0, prs_high))
-
-    return cld_hist
-
-
-# derived_variables is a dictionary to accomodate user specified derived variables.
-# The driver search for available variable keys and functions to calculate derived variable.
-# For example
-# In derived_variable there is an entry for 'PRECT':
-# If 'PRECT' is not available, but both 'PRECC' and 'PRECT' are available in the netcdf variable keys,
-# PRECT is calculated using fuction prect() with precc and precl as inputs.
-
+from typing import Callable, Dict, Tuple
+
+from e3sm_diags.derivations.formulas import (
+    albedo,
+    albedo_srf,
+    albedoc,
+    fldsc,
+    flus,
+    fp_uptake,
+    fsus,
+    lwcf,
+    lwcfsrf,
+    molec_convert_units,
+    netcf2,
+    netcf2srf,
+    netcf4,
+    netcf4srf,
+    netflux4,
+    netflux6,
+    netlw,
+    netsw,
+    pminuse_convert_units,
+    precst,
+    prect,
+    qflx_convert_to_lhflx,
+    qflx_convert_to_lhflx_approxi,
+    qflxconvert_units,
+    restoa,
+    restom,
+    rst,
+    rstcs,
+    swcf,
+    swcfsrf,
+    tauxy,
+    tref_range,
+    w_convert_q,
+)
+from e3sm_diags.derivations.utils import (
+    _apply_land_sea_mask,
+    aplusb,
+    convert_units,
+    cosp_bin_sum,
+    cosp_histogram_standardize,
+    rename,
+)
 
 # A type annotation ordered dictionary that maps a tuple of source variable(s)
 # to a derivation function.
